Add tasks from list view
Now we can add tasks from the list view.
These tasks are not reflected on the DOM yet. The connection
should be done. Also, we need to replace the currentList correctly
with the updated version we have, that has the tasks.

There's also another conflict when adding the tasks to
local storage. This happens at the template.html home.